+++ core/src/main/java/com/trendyol/design/core/dialog/GenericDialogType.kt
@@ -4,10 +4,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.TextStyle
 
-interface StatusDialogType {
+interface GenericDialogType {
 
     @get:Composable
-    val textStyle: TextStyle
+    val textStyle: TextStyle?
 
-    val iconVector: ImageVector
+    val iconVector: ImageVector?
 }