+++ CHANGELOG.md
@@ -132,3 +132,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   file is changed.
 - Used "check" task instead of "detekt" task in CI/CD pipeline to check if everything is OK. "check"
   task includes Detekt and other checks such as running unit tests and Android Lint.
+
+## [Unreleased]
+
+## [0.10.0] - 2024-08-07
+
+### Added
+
+- Introduced `AlertDiag` and `GenericDialog` composables
+- Implemented `TrendyolGenericDialogType` for defining different generic dialog
+  (`Success`, `Error` and `Custom`)
+  of the rating bar.